logger/requirements.txt
requests>=2.5.1
paho-mqtt>=1.5.0
